Add guard for missing browser
When forceKill is run, it seems that it is re-attempted if it fails, which leads to an infinite loop when the browser instance fails to start (and the `this._browser` is undefined). This can happen if playwright installation failed.

With this fix, the infinite loop is avoided, and the launcher fails directly instead.